# backend/openapi.json
{
    "components": {
        "schemas": {
...
            "ExerciseOut": {
                "description": "Exercício como ele chega ao cliente.\n\nA resposta certa **não** faz parte deste schema. Dica e explicação mantêm\nas chaves legadas, mas chegam nulas na leitura inicial; o servidor só as\nlibera pelos fluxos autenticados de tentativa, dica ou revelação.",
                "properties": {
                    "activity_type": {
                        "title": "Activity Type",
                        "type": "string"
                    },
                    "explanation": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Explanation"
                    },
                    "hint": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Hint"
                    },
                    "hint_count": {
                        "title": "Hint Count",
                        "type": "integer"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    },
                    "objective": {
                        "enum": [
                            "recognize",
                            "apply",
                            "correct",
                            "produce",
                            "listen"
                        ],
                        "title": "Objective",
                        "type": "string"
                    },
                    "options": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Options"
                    },
                    "position": {
                        "title": "Position",
                        "type": "integer"
                    },
                    "prompt": {
                        "title": "Prompt",
                        "type": "string"
                    },
                    "skill": {
                        "title": "Skill",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "position",
                    "activity_type",
                    "skill",
                    "objective",
                    "options",
                    "prompt",
                    "hint",
                    "hint_count",
                    "explanation"
                ],
                "title": "ExerciseOut",
                "type": "object"
            },
            "ExerciseWithLessonOut": {
                "properties": {
                    "activity_type": {
                        "title": "Activity Type",
                        "type": "string"
                    },
                    "course_slug": {
                        "title": "Course Slug",
                        "type": "string"
                    },
                    "explanation": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Explanation"
                    },
                    "hint": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Hint"
                    },
                    "hint_count": {
                        "title": "Hint Count",
                        "type": "integer"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    },
                    "lesson_number": {
                        "title": "Lesson Number",
                        "type": "integer"
                    },
                    "objective": {
                        "enum": [
                            "recognize",
                            "apply",
                            "correct",
                            "produce",
                            "listen"
                        ],
                        "title": "Objective",
                        "type": "string"
                    },
                    "options": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Options"
                    },
                    "position": {
                        "title": "Position",
                        "type": "integer"
                    },
                    "prompt": {
                        "title": "Prompt",
                        "type": "string"
                    },
                    "skill": {
                        "title": "Skill",
                        "type": "string"
                    },
                    "unit_slug": {
                        "title": "Unit Slug",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "position",
                    "activity_type",
                    "skill",
                    "objective",
                    "options",
                    "prompt",
                    "hint",
                    "hint_count",
                    "explanation",
                    "course_slug",
                    "unit_slug",
                    "lesson_number"
                ],
                "title": "ExerciseWithLessonOut",
                "type": "object"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "title": "Detail",
                        "type": "array"
                    }
                },
                "title": "HTTPValidationError",
                "type": "object"
            },
...
            "LessonSummaryOut": {
                "description": "O que a trilha de navegação precisa — sem carregar a aula inteira.",
                "properties": {
                    "course_slug": {
                        "title": "Course Slug",
                        "type": "string"
                    },
                    "focus_points": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Focus Points",
                        "type": "array"
                    },
                    "grammar_tag": {
                        "title": "Grammar Tag",
                        "type": "string"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    },
                    "number": {
                        "title": "Number",
                        "type": "integer"
                    },
                    "position": {
                        "title": "Position",
                        "type": "integer"
                    },
                    "slug": {
                        "title": "Slug",
                        "type": "string"
                    },
                    "story_note": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Story Note"
                    },
                    "title": {
                        "title": "Title",
                        "type": "string"
                    },
                    "title_pt": {
                        "title": "Title Pt",
                        "type": "string"
                    },
                    "unit_slug": {
                        "title": "Unit Slug",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "course_slug",
                    "unit_slug",
                    "slug",
                    "position",
                    "number",
                    "title",
                    "title_pt",
                    "grammar_tag",
                    "focus_points",
                    "story_note"
                ],
                "title": "LessonSummaryOut",
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "ctx": {
                        "title": "Context",
                        "type": "object"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "title": "Location",
                        "type": "array"
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                },
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError",
                "type": "object"
            },
...
            "VocabItemWithLessonOut": {
                "properties": {
                    "example_en": {
                        "title": "Example En",
                        "type": "string"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    },
                    "ipa": {
                        "title": "Ipa",
                        "type": "string"
                    },
                    "lesson_number": {
                        "title": "Lesson Number",
                        "type": "integer"
                    },
                    "term": {
                        "title": "Term",
                        "type": "string"
                    },
                    "translation_pt": {
                        "title": "Translation Pt",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "term",
                    "ipa",
                    "translation_pt",
                    "example_en",
                    "lesson_number"
                ],
                "title": "VocabItemWithLessonOut",
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "Aulas de Inglês — API",
        "version": "0.1.0"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/api/courses/{course_slug}/lessons/{number}/exercises": {
            "get": {
                "description": "Lista pública e filtrável; não publica dica, explicação nem gabarito.",
                "operationId": "list_lesson_exercises_api_courses__course_slug__lessons__number__exercises_get",
                "parameters": [
                    {
                        "in": "path",
                        "name": "course_slug",
                        "required": true,
                        "schema": {
                            "title": "Course Slug",
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "number",
                        "required": true,
                        "schema": {
                            "title": "Number",
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "activity_type",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "enum": [
                                        "gap_fill",
                                        "multiple_choice",
                                        "transformation",
                                        "reorder",
                                        "dictation"
                                    ],
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Activity Type"
                        }
                    },
                    {
                        "in": "query",
                        "name": "skill",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "enum": [
                                        "grammar",
                                        "listening"
                                    ],
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Skill"
                        }
                    },
                    {
                        "in": "query",
                        "name": "objective",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "enum": [
                                        "recognize",
                                        "apply",
                                        "correct",
                                        "produce",
                                        "listen"
                                    ],
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Objective"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ExerciseOut"
                                    },
                                    "title": "Response List Lesson Exercises Api Courses  Course Slug  Lessons  Number  Exercises Get",
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "List Lesson Exercises",
                "tags": [
                    "practice"
                ]
            }
        },
...
        "/api/exercises": {
            "get": {
                "description": "Exercícios de uma aula, unidade ou curso.\n\nÉ o que a prova do bloco consome: sem isso a tela teria que baixar as dez\naulas inteiras para montar uma lista de exercícios.",
                "operationId": "listar_exercicios_api_exercises_get",
                "parameters": [
                    {
                        "description": "Filtra por número da aula.",
                        "in": "query",
                        "name": "lesson",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filtra por número da aula.",
                            "title": "Lesson"
                        }
                    },
                    {
                        "description": "Slug do curso; o padrão mantém a consulta no Level 1.",
                        "in": "query",
                        "name": "course",
                        "required": false,
                        "schema": {
                            "default": "voa-level-1",
                            "description": "Slug do curso; o padrão mantém a consulta no Level 1.",
                            "title": "Course",
                            "type": "string"
                        }
                    },
                    {
                        "description": "Limita a avaliação a uma unidade do curso.",
                        "in": "query",
                        "name": "unit",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Limita a avaliação a uma unidade do curso.",
                            "title": "Unit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ExerciseWithLessonOut"
                                    },
                                    "title": "Response Listar Exercicios Api Exercises Get",
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Listar Exercicios",
                "tags": [
                    "lessons"
                ]
            }
        },
...
        "/api/lessons": {
            "get": {
                "description": "Todas as aulas, em ordem, com o mínimo para montar a navegação.",
                "operationId": "listar_aulas_api_lessons_get",
                "parameters": [
                    {
                        "description": "Slug do curso; o padrão mantém o endpoint legado no Level 1.",
                        "in": "query",
                        "name": "course",
                        "required": false,
                        "schema": {
                            "default": "voa-level-1",
                            "description": "Slug do curso; o padrão mantém o endpoint legado no Level 1.",
                            "title": "Course",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/LessonSummaryOut"
                                    },
                                    "title": "Response Listar Aulas Api Lessons Get",
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Listar Aulas",
                "tags": [
                    "lessons"
                ]
            }
        },
        "/api/lessons/{number}": {
            "get": {
                "description": "Uma aula inteira: gramática, frases, vocabulário, pronúncia e exercícios.",
                "operationId": "obter_aula_api_lessons__number__get",
                "parameters": [
                    {
                        "in": "path",
                        "name": "number",
                        "required": true,
                        "schema": {
                            "title": "Number",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LessonDetailOut"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Obter Aula",
                "tags": [
                    "lessons"
                ]
            }
        },
...
        "/api/vocab": {
            "get": {
                "description": "Vocabulário de uma aula, ou de todas. Base do deck de revisão da S4.",
                "operationId": "listar_vocabulario_api_vocab_get",
                "parameters": [
                    {
                        "description": "Filtra por número da aula.",
                        "in": "query",
                        "name": "lesson",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filtra por número da aula.",
                            "title": "Lesson"
                        }
                    },
                    {
                        "description": "Slug do curso; o padrão mantém a consulta no Level 1.",
                        "in": "query",
                        "name": "course",
                        "required": false,
                        "schema": {
                            "default": "voa-level-1",
                            "description": "Slug do curso; o padrão mantém a consulta no Level 1.",
                            "title": "Course",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/VocabItemWithLessonOut"
                                    },
                                    "title": "Response Listar Vocabulario Api Vocab Get",
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Listar Vocabulario",
                "tags": [
                    "lessons"
                ]
            }
        },
...
    }
}